hub: split the replay buffer out of state.rs
No behavior change.
