+++ crates/tribuchet/src/hub/state.rs
@@ -25,6 +25,10 @@ pub(super) const WORKER_GRACE: Duration = Duration::from_secs(30);
 
 use crate::proto::{AttachEvent, BuildRequest, attach_event};
 
+#[path = "state/replay.rs"]
+mod replay;
+pub(super) use replay::Replay;
+
 type EventTx = mpsc::Sender<Result<AttachEvent, Status>>;
 
 /// Cap on the replay buffer of one build. Without it a worker that
@@ -35,118 +39,6 @@ const MAX_REPLAY_BYTES: usize = 256 * 1024 * 1024;
 /// stalled attach client is dropped once it falls this far behind
 /// instead of buffering the whole build a second time.
 const SUB_CHANNEL_SLACK: usize = 1024;
-
-/// Buffered event log of one in-flight build; late identical submissions
-/// (dedupe) replay the buffer and then follow live. The buffer holds the
-/// compressed output chunks too, capped at MAX_REPLAY_BYTES.
-#[derive(Default)]
-pub(super) struct Replay {
-    inner: Mutex<ReplayInner>,
-}
-
-#[derive(Default)]
-struct ReplayInner {
-    events: Vec<AttachEvent>,
-    bytes: usize,
-    /// Buffer cap hit: the backlog is incomplete, so late dedupe
-    /// subscribers must error instead of getting a truncated stream.
-    overflowed: bool,
-    subs: Vec<EventTx>,
-    done: bool,
-}
-
-fn event_size(ev: &attach_event::Event) -> usize {
-    match ev {
-        attach_event::Event::Log(d) => d.len(),
-        attach_event::Event::Output(o) => o.zstd_nar_chunk.len(),
-        attach_event::Event::OutputRestart(p)
-        | attach_event::Event::AddedPath(p)
-        | attach_event::Event::Dispatched(p) => p.len(),
-        attach_event::Event::Error(e) => e.len(),
-        attach_event::Event::ExitCode(_) => 0,
-    }
-    .saturating_add(64)
-}
-
-impl Replay {
-    pub(super) async fn publish(&self, ev: attach_event::Event) {
-        let sz = event_size(&ev);
-        let ev = AttachEvent { event: Some(ev) };
-        let mut inner = self.inner.lock().await;
-        // try_send: a subscriber that stopped reading is dropped (its
-        // attach errors out) instead of buffering unboundedly.
-        inner.subs.retain(|tx| match tx.try_send(Ok(ev.clone())) {
-            Ok(()) => true,
-            Err(mpsc::error::TrySendError::Full(_)) => {
-                tracing::warn!("dropping attach subscriber that fell too far behind");
-                false
-            }
-            Err(mpsc::error::TrySendError::Closed(_)) => false,
-        });
-        if inner.overflowed {
-            return;
-        }
-        if inner.bytes + sz > MAX_REPLAY_BYTES {
-            tracing::warn!("replay buffer cap reached; late dedupe subscribers will be rejected");
-            inner.overflowed = true;
-            inner.events.clear();
-            inner.bytes = 0;
-            return;
-        }
-        inner.bytes += sz;
-        inner.events.push(ev);
-    }
-
-    pub(super) async fn subscribe(&self) -> mpsc::Receiver<Result<AttachEvent, Status>> {
-        let mut inner = self.inner.lock().await;
-        if inner.overflowed {
-            let (tx, rx) = mpsc::channel(1);
-            let _ = tx.try_send(Err(Status::resource_exhausted(
-                "build output exceeded the replay buffer; retry after it finishes",
-            )));
-            return rx;
-        }
-        // Enough capacity for the whole backlog plus live slack (and
-        // one error slot), so the snapshot below cannot drop events.
-        let (tx, rx) = mpsc::channel(inner.events.len() + SUB_CHANNEL_SLACK);
-        for ev in &inner.events {
-            let _ = tx.try_send(Ok(ev.clone()));
-        }
-        if inner.done {
-            // Finished without a verdict in the backlog (e.g. the job
-            // was dropped as abandoned between the dedupe lookup and
-            // this subscribe): an error beats a silently empty stream.
-            let concluded = inner.events.iter().any(|e| {
-                matches!(
-                    e.event,
-                    Some(attach_event::Event::ExitCode(_) | attach_event::Event::Error(_))
-                )
-            });
-            if !concluded {
-                let _ = tx.try_send(Err(Status::unavailable(
-                    "build is no longer in flight; resubmit",
-                )));
-            }
-        } else {
-            inner.subs.push(tx);
-        }
-        rx
-    }
-
-    /// Close all subscriber streams.
-    pub(super) async fn finish(&self) {
-        let mut inner = self.inner.lock().await;
-        inner.done = true;
-        inner.subs.clear();
-    }
-
-    /// Any attach client still listening? Subscribers whose stream was
-    /// dropped count as gone even before publish() prunes them.
-    pub(super) async fn has_subscribers(&self) -> bool {
-        let inner = self.inner.lock().await;
-        inner.subs.iter().any(|tx| !tx.is_closed())
-    }
-}
 
 pub(super) struct Job {
     pub(super) id: String,
